requirements.txt
crewai>=0.11.0
python-dotenv>=1.0.0
resend>=0.6.0
crewai-tools>=0.0.6
